Use 2 more caches for updating documents.
Instead of updating the document once for each transaction Id, queue up all the transaction Ids in the document
that are going to need to be updated, and then use set removal to actually get rid of them.
Also cache knownMissing transaction ids, so we don't have to lookup every transaction that is missing
on every document that is missing it.